+++ my-app/src/lib/stores.js
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const user = writable()
