+++ my-app/src/routes/[userid]/+page.server.js
@@ -1,0 +1,6 @@
+export function load({params}) {
+    return {
+        userid: params.userid,
+    }
+
+}
