+++ my-app/src/routes/[userid]/+page.svelte
@@ -1,0 +1,5 @@
+This is where my User homepage is.
+<script>
+    export let data;
+    console.log(data);
+</script>
