+++ my-app/src/routes/[userid]/characters/+page.svelte
@@ -1,0 +1,1 @@
+This is where my list of characters goes.
