+++ my-app/src/routes/[userid]/characters/[characterid]/+page.svelte
@@ -1,0 +1,1 @@
+This is where a character will go.
