+++ my-app/src/routes/[userid]/characters/[characterid]/charcreator/2/+page.svelte
@@ -11,14 +11,7 @@
 		<a class="pageLink" href="../Char">Characters</a>
 	</div>
 </div>
-<div class="nextBack">
-	<div class="back">
-		<a class="nextBackLink" href="charcreator1">Back</a>
-	</div>
-	<div class="next">
-		<a class="nextBackLink" href="charcreator3">Next</a>
-	</div>
-</div>
+
 <div class="wrapper">
 	<div class="skill">
 		<label class="label" for="athletics"
@@ -188,35 +181,7 @@
 		color: cyan;
 	}
 
-	.nextBack {
-		width: 100vw;
-		height: 5em;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		background-color: #999aad;
-	}
-
-	.back,
-	.next {
-		display: flex;
-		border: solid black 0.5px;
-		justify-content: center;
-		align-items: center;
-		height: 100%;
-		width: 50vw;
-		font-size: 40px;
-	}
-
-	.nextBackLink {
-		text-decoration: none;
-		color: black;
-	}
-
-	.nextBackLink:hover {
-		text-decoration: underline;
-		color: cyan;
-	}
+	
 
 	.skill {
 		display: flex;
