+++ my-app/src/routes/[userid]/characters/[characterid]/charcreator/5/+page.svelte
